Divided databinding code
