+++ app/src/main/java/in/hangang/hangang/ui/MainActivity.kt
@@ -1,6 +1,6 @@
 package `in`.hangang.hangang.ui
 
-import `in`.hangang.core.ActivityBase
+import `in`.hangang.core.base.activity.ActivityBase
 import `in`.hangang.hangang.R
 import android.os.Bundle
 import androidx.navigation.Navigation
