+++ core/src/main/res/layout/activity_base.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityBase">
+    tools:context=".base.activity.ActivityBase">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
